server/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pymongo>=4.6.0
python-dotenv>=1.0.0
passlib[bcrypt]>=1.7.4
authlib>=1.3.0
starlette>=0.27.0
httpx>=0.25.0
itsdangerous>=2.1.2
pyjwt>=2.8.0
bcrypt>=4.1.2
email-validator>=2.1.0
groq>=0.4.1
requests>=2.31.0
mcp[cli]>=1.0.0
redis>=5.0.1
selenium>=4.15.0
numpy>=1.24.0
pinecone-client>=3.0.0
gradio_client>=0.8.0
langgraph>=0.2.16
langchain>=0.2.14
langchain-groq>=0.1.9
langchain-community>=0.2.12
langchain-core>=0.2.38
langsmith>=0.1.77